moved navbar to cmpnts/nav/
